实现一个基础的ext2文件系统

+++ kernel/src/mm/bitmap.h
@@ -4,7 +4,7 @@
 
 typedef struct
 {
-    uint8_t *buffer;
+    uint64_t *buffer;
     size_t length;
 } Bitmap;
